src/main/java/edu/yctc/genesis/entity/ClassDO.java
package edu.yctc.genesis.entity;

import java.util.Date;

public class ClassDO {

    /** 主键 */
    private long id;
    /** 学院id */
    private long academyId;
    /** 专业id */
    private long professionId;
    /** 班级号 */
    private String number;
    /** 创建时间 */
    private Date createTime;
    /** 修改时间 */
    private Date modifyTime;

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getNumber() {
        return number;
    }

    public void setNumber(String number) {
        this.number = number;
    }

    public Date getCreateTime() {
        return createTime;
    }

    public void setCreateTime(Date createTime) {
        this.createTime = createTime;
    }

    public Date getModifyTime() {
        return modifyTime;
    }

    public void setModifyTime(Date modifyTime) {
        this.modifyTime = modifyTime;
    }

    public long getAcademyId() {
        return academyId;
    }

    public void setAcademyId(long academyId) {
        this.academyId = academyId;
    }

    public long getProfessionId() {
        return professionId;
    }

    public void setProfessionId(long professionId) {
        this.professionId = professionId;
    }

    @Override
    public String toString() {
        StringBuilder builder = new StringBuilder();
        builder.append("ClassDO [id=");
        builder.append(id);
        builder.append(", academyId=");
        builder.append(academyId);
        builder.append(", professionId=");
        builder.append(professionId);
        builder.append(", number=");
        builder.append(number);
        builder.append(", createTime=");
        builder.append(createTime);
        builder.append(", modifyTime=");
        builder.append(modifyTime);
        builder.append("]");
        return builder.toString();
    }

}
